Move and reuse environment_variable_helper
