+++ source/core/src/core/silver/entry_points.py
@@ -1,5 +1,5 @@
 import core.silver.application.streams.calculated_stream as calculated_stream
-from core.silver.infrastructure.helpers.environment_variable_helper import get_applicationinsights_connection_string
+from core.utility.environment_variable_helper import get_applicationinsights_connection_string
 
 
 def stream_calculated_measurements() -> None:
