+++ source/core/src/core/utility/shared_helpers.py
@@ -1,7 +1,4 @@
 # This will contain methods that should be put into a 3rd party library, and reused across multiple projects.
-import os
-from enum import Enum
-from typing import Any
 
 
 def get_storage_base_path(
@@ -21,15 +18,3 @@ def get_checkpoint_path(
 
 def get_full_table_name(database: str, table: str) -> str:
     return f"{database}.{table}"
-
-
-class EnvironmentVariable(Enum):
-    DATALAKE_STORAGE_ACCOUNT = "DATALAKE_STORAGE_ACCOUNT"
-
-
-def get_env_variable_or_throw(variable: EnvironmentVariable) -> Any:
-    env_variable = os.getenv(variable.name)
-    if env_variable is None:
-        raise ValueError(f"Environment variable not found: {variable.name}")
-
-    return env_variable
